refactor(api): generate 관련 코드를 generateClient.ts로 분리
missionClient.ts에 혼재되어 있던 GenerateRequest·GenerateResponse·
GenerateStatusResponse 타입과 generate·getGenerateStatus 함수를
generateClient.ts로 이동.
missionClient.ts는 미션 조회 전용으로 정리.
